lib/osml-apis/cdk/lambda/authorizer/requirements.txt
PyJWT[crypto]==2.8.0
requests==2.31.0
